refactor: split hooks into individual files
